# ErsatzTV/wwwroot/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ErsatzTV | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/libraries": {
      "get": {
        "tags": [
          "Libraries"
        ],
        "summary": "Get all configured libraries",
        "operationId": "GetLibraries",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LibraryViewModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LibraryViewModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LibraryViewModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Libraries"
        ],
        "summary": "Create a local library",
        "operationId": "CreateLocalLibrary",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocalLibraryRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocalLibraryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocalLibraryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocalLibraryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LocalLibraryViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalLibraryViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalLibraryViewModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/shows": {
      "get": {
        "tags": [
          "Libraries"
        ],
        "summary": "Get all television shows",
        "operationId": "GetTelevisionShows",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NamedMediaItemViewModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NamedMediaItemViewModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NamedMediaItemViewModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{id}/scan": {
      "post": {
        "tags": [
          "Libraries"
        ],
        "summary": "Scan library",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{id}/scan-show": {
      "post": {
        "tags": [
          "Libraries"
        ],
        "summary": "Scan show",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ScanShowRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanShowRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanShowRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScanShowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLocalLibraryRequest": {
        "required": [
          "name",
          "mediaKind",
          "paths"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "mediaKind": {
            "$ref": "#/components/schemas/LibraryMediaKind"
          },
          "paths": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "LibraryMediaKind": {
        "type": "integer"
      },
      "LibraryViewModel": {
        "required": [
          "libraryKind",
          "id",
          "name",
          "mediaKind",
          "mediaSourceId",
          "mediaSourceName"
        ],
        "type": "object",
        "properties": {
          "libraryKind": {
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "mediaKind": {
            "$ref": "#/components/schemas/LibraryMediaKind"
          },
          "mediaSourceId": {
            "type": "integer",
            "format": "int32"
          },
          "mediaSourceName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LocalLibraryViewModel": {
        "required": [
          "id",
          "name",
          "mediaKind",
          "mediaSourceId"
        ],
        "type": "object",
        "properties": {
          "libraryKind": {
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "mediaKind": {
            "$ref": "#/components/schemas/LibraryMediaKind"
          },
          "mediaSourceId": {
            "type": "integer",
            "format": "int32"
          },
          "mediaSourceName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "NamedMediaItemViewModel": {
        "required": [
          "mediaItemId",
          "name"
        ],
        "type": "object",
        "properties": {
          "mediaItemId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ScanShowRequest": {
        "required": [
          "showTitle"
        ],
        "type": "object",
        "properties": {
          "showTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "deepScan": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Only enforced when the ETV_API_KEY environment variable is set; the api is anonymous otherwise.",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ],
...
}